Ok auth password now can be optionally set, refactored a lot of IQ calls into IqComponentModel, and fun
